+++ NEWS.md
@@ -1,3 +1,10 @@
+## BayesianMCPMod 1.0.1 (03-Apr-2024)
+
+- Re-submission of the 'BayesianMCPMod' package
+- Removed a test that occasionally failed on the fedora CRAN test system
+- Fixed a bug that would return wrong bootstrapped quantiles in getBootstrappedQuantiles()
+- Added getBootstrapSamples(), a separate function for bootstrapping samples
+
 ## BayesianMCPMod 1.0.0 (31-Dec-2023)
 
 - Initial release of the 'BayesianMCPMod' package